create a hook with battle logic
